+++ CMakeLists.txt
@@ -143,7 +143,7 @@ set(SOURCES
     src/capture/screenshot_window.cpp
     src/config/config_service.cpp
     src/config/ui/config_ui.cpp
-    src/config/ui/generic_section_renderer.cpp
+    src/config/ui/section_renderer.cpp
     src/keybinds/input_manager.cpp
     src/keybinds/keybind_manager.cpp
     src/render/asset_manager.cpp
